Simplify usage of groupButtons
Instead of having to manually find all siblings, automagically group buttons by just specifying one of them.
